feat(api): 타임존 및 언어 쿼리 스키마 추가 및 api 적용
