+++ packages/api/src/domains/analysis/index.ts
@@ -1,5 +1,4 @@
 export * from "./category-analysis.schema";
-export * from "./enum.schema";
 export * from "./frequently-visited-websites.schema";
 export * from "./longest-stayed-website.schema";
 export * from "./screen-time.schema";
