+++ packages/api/src/domains/user/user-profile.schema.ts
@@ -1,13 +1,15 @@
 import { z } from "zod";
 
-import { CreateResponseSchema } from "../../schema";
+import { CreateResponseSchema } from "../../schemas";
+import { LanguageSchema, TimeZoneSchema } from "../../schemas/enum.schema";
 
 const UserProfileSchema = z.object({
   email: z.string(),
   firstName: z.string(),
   lastName: z.string(),
   imageUrl: z.string(),
-  timeZone: z.string().optional(),
+  timeZone: TimeZoneSchema.optional(),
+  language: LanguageSchema.optional(),
   recapPeriod: z.string().nullable().optional(),
   excludedDomains: z.array(z.string()),
 });
