+++ packages/api/src/services/recap-api.service.ts
@@ -1,6 +1,6 @@
-import type { DateQueryType } from "../domains/analysis/enum.schema";
 import { GetRecapResponseSchema } from "../domains/recap/generate-recap.schema";
 import type { RestAPIProtocol } from "../rest/types";
+import type { DateQueryType } from "../schemas/enum.schema";
 
 export class RecapAPIService {
   constructor(private fetch: RestAPIProtocol) {}
